==> s1/1_business_understanding.ipynb <==
# business understanding
------------

This initial phase focuses on understanding the project objectives and requirements from a business perspective, then converting this knowledge into a data mining problem definition and a preliminary plan designed to achieve the objectives.


## determine business objectives
----------

### task

The first objective of the data analyst is to thoroughly understand, from a business perspective, what the client really wants to accomplish. Often the client has many competing objectives and constraints that must be properly balanced. The analyst's goal is to uncover important factors, at the beginning, that can influence the outcome of the project. A possible consequence of neglecting this step is to expend a great deal of effort producing the right answers to the wrong questions.

### output

#### background

Record the information that is known about the organization's business situation at the beginning of the project.

#### business objectives 

Describe the customer's primary objective, from a business perspective. In addition to the primary business objective, there are typically other related business questions that the customer would like to address. For example, the primary business goal might be to keep current customers by predicting when they are prone to move to a competitor. Examples of related business questions are "How does the primary channel (e.g., ATM, visit branch, internet) a bank customer uses affect whether they stay or go?" or "Will lower ATM fees significantly reduce the number of high-value customers who leave?"

#### business success criteria 

Describe the criteria for a successful or useful outcome to the project from the business point of view. This might be quite specific and able to be measured objectively, such as reduction of customer churn to a certain level or general and subjective such as "give useful insights into the relationships." In the latter case it should be indicated who makes the subjective judgment.


## assess situation
-------------

### task

This task involves more detailed fact-finding about all of the resources,constraints, assumptions and other factors that should be considered in determining the data analysis goal and project plan. In the previous task, your objective is to quickly get to the crux of the situation. Here, you want to flesh out the details.

### output

#### inventory of resources

List the resources available to the project, including: personnel (business experts, data experts, technical support, data mining personnel), data (fixed extracts, access to live warehoused or operational data), computing resources (hardware platforms) and software (data miningtools, other relevant software).

#### requirements, assumptions and constraints

List all requirements of the project including schedule of completion, comprehensibility and quality of results and security as well as legal issues.As part of this output, make sure that you are allowed to use the data. List the assumptions made by the project. 

These may be assumptions about the data that can be checked during data mining, but may also include non-checkable assumptions about the business upon which the project rests. It is particularly important to list the latter if they form conditions on the validity of the results.

List the constraints on the project. These may be constraints on the availability of resources, but may also include technological constraints such as the size of data that it is practical to use for modeling.

#### Risks and contingencies 

List the risks or events that might occur to delay the project or cause it to fail. List the corresponding contingency plans; what action will be taken if the risks happen.

#### Terminology

Compile a glossary of terminology relevant to the project. This may include two components: 
(1) A glossary of relevant business terminology, which forms part of the business understanding available to the project. Constructing this glossary is a useful "knowledge elicitation" and education exercise.
(2) A glossary of data mining terminology, illustrated with examples relevant to the business problem in question.

#### Costs and benefits

Construct a cost-benefit analysis for the project, which compares the costs of the project with the potential benefit to the business if it is successful. The comparison should be as specific as possible, for example using monetary measures in a commercial situation.


## determine data mining goals
-----------

### task

A business goal states objectives in business terminology. A data mining goal states project objectives in technical terms. For example, the business goal might be "Increase catalog sales to existing customers." A data mining goal might be "Predict how many widgets a customer will buy, given their purchases over the past three years, demographic information (age, salary, city, etc.) and the price of the item."

### output

#### data mining goals

Describe the intended outputs of the project that enables the achievement of the business objectives.

#### data mining success criteria

Define the criteria for a successful outcome to the project in technical terms, for example a certain level of predictive accuracy or a propensity to purchase profile with a given degree of "lift." As with business success criteria, it may be necessary to describe these in subjective terms, in which case the person or persons making the subjective judgment should be identified.


## produce project plan
----------------

### task

Describe the intended plan for achieving the data mining goals and thereby achieving the business goals. The plan should specify the anticipated set of steps to be performed during the rest of the project including an initial selection of tools and techniques.

### output

#### project plan

List the stages to be executed in the project, together with duration, resources required, inputs, outputs and dependencies. Where possible make explicit the large-scale iterations in the data mining process, for example repetitions of the modeling and evaluation phases. As part of the project plan, it is also important to analyze dependencies between time schedule and risks. Mark results of these analyses explicitly in the project plan, ideally with actions and recommendations if the risks appear.

Note: the project plan contains detailed plans for each phase. For example, decide at this point which evaluation strategy will be used in the evaluation phase. The project plan is a dynamic document in the sense that at the end of each phase a review of progress and achievements is necessary and an update of the project plan accordingly is recommended. Specific review points for these reviews are part of the project plan, too.

#### initial assessment of tools and techniques

At the end of the first phase, the project also performs an initial assessment of tools and techniques. Here, you select a data mining tool that supports various methods for different stages of the process, for example. It is important to assess tools and techniques early in the process since the selection of tools and techniques possibly influences the entire project.


# Answers


## Determine business objectives
##### Background:
Organization have information about products: their name, available quantity, price, country where a product was created, time of a purchase, id of a customer(I guess customer means company\individual that sold products to organization).
##### Business objective
Business objective is to gain maximum profit. To do it we can minimize cost to prepare products(e.g. minimize traveling expenses), and improve sales by selling in more places for which we need to find out where we need to add another stores or pulling out stores with low sales or finding duplicate products and using most profitable product out of these for sale.
##### business success criteria
Let's say success of a project will be if we: minimize traveling costs by 5%, increasing available customers by 10%, reduce duplicates sales.
## Assess situation
###### inventory of resources
4 people: 1 Business expert, 1 data mining expert, 2 ml engineers. 4 servers(all with medium cost GPUs and medium cost CPU(for example at AWS\organization), 32GB memory, and enough disk for storing data + tools): 3 for developers, 1 for storing and calculating weights for a last version of a model. All data should be persisted on all servers using SQL database(for example PostgreSQL). ML engineers servers should have some flavour of linux operating system(for example ubuntu with necessary libs i.e. wget, curl e.t.c.) with tools for machine learning(python, pytorch, conda, jupyter-notebook, pycharm e.t.c.), data mining expert should have same flavour of linux with installed data mining tools(python, numpy, pandas, sklearn, pycharm). Last server should have everything needed for calculating final model(same as ML engineer server + some of data mining tools). Servers should be configred in a network with secure gateway(password + changed ssh port + firewall\security groups configured to access only developers of a project).
###### requirements, assumptions and constraints
Project should be completed by 3 months. Results should be archivable by a cost in such a way that after proposed changes, cost will be retractable by 3 year. All people working on project should sign NDA. Data should be collected with customer agreements. Data after data mining can be checked, at least by ids(company names can be hidden). Data size should be at least 5 GB and utmost 1TB.
###### Risks and contingencies
If technical issues occur(e.g. power outage of servers) organization should provide new hardware utmost in a week after issue. If health issues of developers appear - in case of inability to work organization should hire another developer. In case of needed extended time project can be extended by utmost month.
###### Terminology
All terminology for Machine learning + data mining (example: http://twocrows.com/data-mining/dm-glossary/)
And terminology of a dataset.
###### Costs and benefits
Results should be archivable by cost in such a way that after proposed changes, cost will be retractable by 3 year.

## Determine data mining goals
##### data mining goals
Finding out new places for stores, stores that should be pulled out, which customers have non-unique products. Other solutions for maximizing profits.

##### data mining success criteria
All goals should have fullfilled. Requirement about new stores and pulling out stores is scored by available customers who will visit stores and sales, after changes number of customers should increase at least by 5%.

## produce project plan
###### project plan
0 Finding team for a project\
1 Introduction to problem\
2 Finding new solutions for maximizing profit\
3 Developing solution based on stage 2 and proposed solutions\
4 Calculating final model\
5 Testing it on large dataset\
6 Presenting final model\
7 Make changes and test model on real world
###### initial assessment of tools and techniques
ML: Choose techniques\algorithms which are more accurate.\
Data Minning: take best possible solutions out of all proposed ones.
