==> s1/3_data_preparation.ipynb <==
# # mount drive
# from google.colab import drive
# drive.mount('/content/drive')

# # move directory
# import os
# colab_dir = "./drive/MyDrive/"
# os.chdir(colab_dir)

# import libraries
!pip install tabulate
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
import matplotlib_inline
%matplotlib inline

# set random seed
import random
random.seed(335)

# magic word
pd.set_option('display.max_rows', 500)
pd.set_option('display.max_columns', 500)

# for better viz
import pprint
import warnings
warnings.filterwarnings('ignore')

### reference
-------------------

- [pandas cheat sheet](https://github.com/pandas-dev/pandas/tree/master/doc/cheatsheet)
- [numpy cheat sheet(data camp)](https://www.datacamp.com/community/blog/python-numpy-cheat-sheet)
- [scikit-learn cheat sheet(data camp)](datacamp.com/community/blog/scikit-learn-cheat-sheet)

# data preparation
---------------------
The data preparation phase covers all activities to construct the final dataset (data that will be fed into the modeling tool(s)) from the initial raw data. Data preparation tasks are likely to be performed multiple times and not in any prescribed order. Tasks include table, record and attribute selection as well as transformation and cleaning of data for modeling tools.

## select data
----------

### task

Decide on the data to be used for analysis. Criteria include relevance to the data mining goals, quality and technical constraints such as limits on data volume or data types. Note that data selection covers selection of attributes (columns) as well as selection of records (rows) in a table.

### output

List the data to be included/excluded and the reasons for these decisions.

#### select data source
| # | data | included/excluded | reasons | quality | volume/data types |
|:---:|:---|:---|:---|:---|:---|
| 1 | Retail | included | it quantifies sales | bad | excel table |

#### select attributes & records

| # | data | included/excluded | reasons | quality | volume/data types |
|:---:|:---|:---|:---|:---|:---|
| 1 | Quantity | included | it quantifies sales | bad | int |
| 2 | InvoiceDate | included | can be used to see patterns| medium | date |
| 3 | UnitPrice | included | it quantifies sales | bad | double |
| 4 | Country | included | signify price for traveling goods| good | string |
| 5 | CustomerID | included | used as index, give us information that customer was the same as with other product | good | int |
| 6 | StockCode | excluded | used as index, doesn't show any information | good | string |
| 7 | InvoiceNo | excluded | used as index, doesn't show any information + missing in almost all rows | very bad | int |
| 8 | Description | included | products with similar words can be used for another customers | good | string |


## clean data
------------------

### task

Raise the data quality to the level required by the selected analysistechniques. This may involve selection of clean subsets of the data, the insertion of suitable defaults or more ambitious techniques such as the estimation of missing data by modeling.

### output

Describe what decisions and actions were taken to address the data quality problems reported during the verify data quality task of the data understanding phase. Transformations of the data for cleaning purposes and the possible impact on the analysis results should be considered.


path = "data/"
file_name = "Online Retail.xlsx"
sheet_name = "Online Retail"
df = pd.read_excel(path + file_name, sheet_name=sheet_name)
df

# remove unnncesary columns
columns = ['StockCode', 'InvoiceNo']
df = df.drop(columns=columns)

del_flg1 = df['Quantity'] <= 0
def_neg = df.copy()
def_neg = def_neg[del_flg1]
def_neg["sum"] = def_neg.apply(lambda row: row['Quantity']*row['UnitPrice'], axis=1)
def_neg.groupby("Country")["sum"].sum()
#price of all returned items from country

countries = pd.get_dummies(def_neg["Country"], dummy_na=False)
def_neg = def_neg.drop('Country', axis = 1, errors='ignore')
def_neg['InvoiceDate'] = pd.to_datetime(def_neg['InvoiceDate'])
def_neg['time'] = def_neg['InvoiceDate'].dt.hour
# Join the encoded df
def_neg = def_neg.join(countries)
def_neg

def_neg.to_csv(path + "negative.xlsx", index=False)

# clean outlier
h = df["UnitPrice"].quantile(0.99)
l = df["UnitPrice"].quantile(0.1)
df = df[(df["UnitPrice"] < h) & (df["UnitPrice"] > l)]
## clean minus data
column1 = 'Quantity'
column2 = 'UnitPrice'
del_flg1 = df[column1] <= 0
del_flg2 = df[column2] <= 0
df = df.drop(df[del_flg1].index)
df = df.drop(df[del_flg2].index)

# clean duplicate records
df = df.drop_duplicates()

df

## construct data
----------

### task

This task includes constructive data preparation operations such as the production of derived attributes, entire new records or transformed values for existing attributes.

### output

#### derived attributes

Derived attributes are new attributes that are constructed from one or more existing attributes in the same record. 

Examples: area = length * width

#### generated records

Describe the creation of completely new records. 

Example: create records for customers who made no purchase during the past year.There was no reason to have such records in the raw data, but for modeling purposes it might make sense to explicitly represent the fact that certain customers made zero purchases.


# derive attributes
df["sum"] = df.apply(lambda row: row['Quantity']*row['UnitPrice'], axis=1)
#price of all items for every item
df

df.groupby("Country")["sum"].sum()
#price of all items from country

import datetime as dt

#added time column for time of the day of the selling of product
df['InvoiceDate'] = pd.to_datetime(df['InvoiceDate'])
df['time'] = df['InvoiceDate'].dt.hour
#change date to ordinal numbers
df['InvoiceDate'] = df['InvoiceDate'].map(dt.datetime.toordinal)

#hot encoding of the contries
countries = pd.get_dummies(df["Country"], dummy_na=False)
df = df.drop('Country', axis = 1, errors='ignore')
# Join the encoded df
df = df.join(countries)
df = df[df["Unspecified"]==0]
df = df.drop(["Unspecified"], axis=1)

#cleaning na values
df = df[df['CustomerID'].notna()]

df.to_csv(path + "input.xlsx", index=False)

# generate records

## integrate data
-------------

### task

These are methods whereby information is combined from multiple tables or records to create new records or values.

### output

Merging tables refers to joining together two or more tables that have different information about the same objects. 

Example: a retail chainhas one table with information about each store's general characteristics(e.g., floor space, type of mall), another table with summarized sales data (e.g., profit, percent change in sales from previous year) and another with information about the demographics of the surrounding area. Each of these tables contains one record for each store. These tables can be merged together into a new table with one record foreach store, combining fields from the source tables.

Merged data also covers aggregations. Aggregation refers to operations where new values are computed by summarizing together information from multiple records and/or tables. For example, converting a table ofcustomer purchases where there is one record for each purchase into a new table where there is one record for each customer, with fields such as number of purchases, average purchase amount, percent of orders charged to credit card, percent of items under promotion, etc.


# format data
----------------

### task

Formatting transformations refer to primarily syntactic modifications made to the data that do not change its meaning, but might be required by the modeling tool.

### output

Some tools have requirements on the order of the attributes, such as the first field being a unique identifier for each record or the last field being the outcome field the model is to predict.

It might be important to change the order of the records in the dataset. Perhaps the modeling tool requires that the records be sorted according to the value of the outcome attribute. A common situation is that the records of the dataset are initially ordered in some way but the modeling algorithm needs them to be in a fairly random order. For example, when using neural networks it is generally best for the records to be presented in a random order although some tools handle this automatically with-out explicit user intervention.

Additionally, there are purely syntactic changes made to satisfy the requirements of the specific modeling tool. 

Examples: removing commas from within text fields in comma-delimited data files, trimming all values to a maximum of 32 characters.


## note/questions
-------------

#### select data
Quantity, InvoiceDate, UnitPrice, Country are selected columns. \
StockCode, CustomerID, InvoiceNo, Description are dropped.
#### clean data
rows where Quantity and UnitPrice is negative\zero are dropped, duplicates are dropped too.
#### construct data
Constructed data is:
 - Price of all items for every item
 - Time of the day of sell
 - Price of all items for every country

I didn't generate new rows, because it will make quality of data even lower, as generated data will be random.
#### integrate data
We don't need to integrate data, because we have only 1 table.
#### format data
We don't need to sort\rename data. But InvoiceDate can be converted to ordinal number
