==> s1/4_modeling.ipynb <==
# # mount drive
# from google.colab import drive
# drive.mount('/content/drive')

# # move directory
# import os
# colab_dir = "./drive/MyDrive/"
# os.chdir(colab_dir)

# import libraries
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
import matplotlib_inline
%matplotlib inline

# # import python module
# from python.xxxx import XXXX

# set random seed
import random
random.seed(335)

# magic word
pd.set_option('display.max_rows', 500)
pd.set_option('display.max_columns', 500)

# for better viz
import pprint
import warnings
warnings.filterwarnings('ignore')

### reference
-------------------

- [pandas cheat sheet](https://github.com/pandas-dev/pandas/tree/master/doc/cheatsheet)
- [numpy cheat sheet(data camp)](https://www.datacamp.com/community/blog/python-numpy-cheat-sheet)
- [scikit-learn cheat sheet(data camp)](datacamp.com/community/blog/scikit-learn-cheat-sheet)

# modeling
---------------------
In this phase, various modeling techniques are selected and applied and their parameters are calibrated to optimal values. Typically, there are several techniques for the same data mining problem type. Some techniques have specific requirements on the form of data. Therefore, stepping back to the data preparation phase is often necessary.

## select modeling techuique
----------

### task

As the first step in modeling, select the actual modeling technique that is to be used. Whereas you possibly already selected a tool in business understanding, this task refers to the specific modeling technique, e.g.,decision tree building with C4.5 or neural network generation with back propagation. If multiple techniques are applied, perform this task for each technique separately.

### output

#### modeling technique

Document the actual modeling technique that is to be used.

#### modeling assumptions

Many modeling techniques make specific assumptions on the data, e.g.,all attributes have uniform distributions, no missing values allowed, class attribute must be symbolic etc. Record any such assumptions made.

## generate test design
------------------

### task

Before we actually build a model, we need to generate a procedure or mechanism to test the model's quality and validity. For example, in supervised data mining tasks such as classification, it is common to use error rates as quality measures for data mining models. Therefore, we typically separate the dataset into train and test set, build the model on the train set and estimate its quality on the separate test set.

### output

Describe the intended plan for training, testing and evaluating the models. A primary component of the plan is to decide how to divide the available dataset into training data, test data and validation datasets.


## build model
----------

### task

Run the modeling tool on the prepared dataset to create one or more models.

### output

#### parameter settings 

With any modeling tool, there are often a large number of parameters that can be adjusted. List the parameters and their chosen value, along with the rationale for the choice of parameter settings. 

#### models 

These are the actual models produced by the modeling tool, not a report.

#### model description

describe the resultant model. Report on the interpretation of the models and document any difficulties encountered with their meanings.


df = pd.read_csv("data/input.xlsx")
df

X = df.drop(["sum", "Description", 'InvoiceDate', 'CustomerID'], axis=1)
y = df["sum"]
X

# split data
## random split
from sklearn.model_selection import train_test_split
X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=0, test_size=0.3)

### regression
-------------

from sklearn.linear_model import LinearRegression
regression = LinearRegression().fit(X_train, y_train)

for i in range(len(X_train.columns)):
    print(f"type {X_train.columns[i]} has coef {regression.coef_[i]}")
print("\nThese countries has negative coefficients:")
for i in range(len(X_train.columns[4:])):
    coef = regression.coef_[4:][i]
    if coef < 0:
        print(X_train.columns[4:][i])
#These countries are profitable, because negative coefficient means other coefficients matter more(quantity, price, time)
print("\nThese countries has large coefficients more than 5:")
for i in range(len(X_train.columns[4:])):
    coef = regression.coef_[4:][i]
    if coef > 5:
        print(X_train.columns[4:][i])
#it isn't profitable to sell there, it means an organization should demolish more stores there, also logistics for these countries are costly.

df.groupby('time').sum()["sum"]
#best possible time to sell is 12:00

df_neg = pd.read_csv("data/negative.xlsx")
df_neg = df_neg.drop(["Description", "InvoiceDate", "CustomerID"], axis=1)
df_neg

from sklearn.linear_model import LinearRegression
y = df_neg["sum"]
x = df_neg.drop(["sum"], axis=1)
regression_neg = LinearRegression().fit(x,y)

for i in range(len(x.columns)):
    coef = regression_neg.coef_[i]
    print(f"{x.columns[i]}: {coef}")
    #more coef more likely customer will give back product

### classification
------------

from sklearn.cluster import KMeans
from sklearn.metrics import silhouette_score
kmeans = KMeans().fit(X_train)
centers = kmeans.cluster_centers_
centers

silhouette_score(X_test, kmeans.predict(X_test))

pd.options.display.float_format = '{:.4f}'.format
pd.DataFrame(data=centers, columns=X_train.columns)

#### imbalanced data

# !pip install -U scikit-learn
# !pip install -U imbalanced-learn
#if you haven't installed. Reboot jupyter after installing packages, if you have errors with imports.

# imbalanced data:
#https://www.analyticsvidhya.com/blog/2020/07/10-techniques-to-deal-with-class-imbalance-in-machine-learning/
## random under sampler
from imblearn.under_sampling import RandomUnderSampler
sampler = RandomUnderSampler()

## random over sampler
# from imblearn.over_sampling import RandomOverSampler
# sampler = RandomOverSampler()

## SMOTE
# from imblearn.over_sampling import SMOTE
# sampler = SMOTE()

## TomekLinks
# from imblearn.under_sampling import TomekLinks
# sampler = TomekLinks()

## NearMiss
# from imblearn.under_sampling import NearMiss
# sampler = NearMiss()

## sampling
# X_train, y_train = sampler.fit_resample(X_train, y_train)

#Don't needed because k means is unsupervised learning, regression doesn't predict categorical data.

#### modeling

# # SVM
# from sklearn.svm import SVC
# model = SVC(kernel='linear', random_state=None)
# model.fit(X_train, y_train)
# y_pred = model.predict(X_test)

# # LightGBM
# !pip install optuna
# import optuna.integration.lightgbm as lgb
# lgb_train = lgb.Dataset(X_train, y_train)
# lgb_eval = lgb.Dataset(X_test, y_test, reference=lgb_train)
# params = {
#     "objective" : "multiclass",
#     "metric" : "multi_logloss",
#     "num_class" : len(y.unique())
# }
# model = lgb.train(params, lgb_train, valid_sets=lgb_eval)
# y_prob = model.predict(X_test, num_iteration=model.best_iteration)
# y_pred = np.argmax(y_prob, axis=1)

## assess model
-------------

### task

The data mining engineer interprets the models according to his domain knowledge, the data mining success criteria and the desired test design. This task interferes with the subsequent evaluation phase. Whereas the data mining engineer judges the success of the application of modeling and discovery techniques more technically, he contacts business analysts and domain experts later in order to discuss the data mining results in the business context. Moreover, this task only considers models whereas the evaluation phase also takes into account all other results that were produced in the course of the project. The data mining engineer tries to rank the models. He assesses the models according to the evaluation criteria. As far as possible he also takes into account business objectives and business success criteria. In most data mining projects, the data mining engineer applies a single technique more than once or generates data mining results with different alternative techniques. In this task, he also compares all results according to the evaluation criteria.

### output

#### model assessment

Summarize results of this task, list qualities of generated models (e.g.,in terms of accuracy) and rank their quality in relation to each other. 

#### revised parameter settings

According to the model assessment, revise parameter settings and tune them for the next run in the Build Model task. Iterate model building and assessment until you strongly believe that you found the best model(s). Document all such revisions and assessments.


### regression
-----------------

from sklearn.metrics import mean_squared_error
y_pred = regression.predict(X_test)
mse = mean_squared_error(y_test, y_pred)
count = X_test.shape[0]
print(f'for {count} rows mse is {mse}')

### classification
----------------


# # accuracy
# from sklearn.metrics import accuracy_score
# accuracy = accuracy_score(y_test, y_pred)
# print(str('{:.1g}'.format(accuracy * 100)) + '%')

# # confusion matrix
# from sklearn.metrics import confusion_matrix
# cm = confusion_matrix(y_test, y_pred)
# print(cm)

#### binary classification

#### multi-class classification

## note/questions
-------------

#### select modeling technique
modeling technique - linear regression. \
Modeling assumptions - no missing values allowed, all attributes have uniform distributions
#### generate test design
Dataset is divided to 30% test set, 70% - train set.
#### build model
Regression: Model is build to determine profit based of in which country customer is located, price of product, date of purchase. So that we can identify which products(from which countries and their price and quantity) are best suited for profit and in which countries we should sell products and which time is best for selling and should we increase quantity of the same products.
Another model is about chances to give product back, based on characteristics such as countries, time, quantity and price.
#### assess model
For test set(103056 rows) mse was around 1770. \
In average model has 0.017 error. \
Model doesn't have any hyperparameters, so no tuning needed.
We won't need to assess another model, because we just determining chances to give product back based on characteristics, we aren't predicting anything, it is used for this data.
