==> s1/5_evaluation.ipynb <==
# import libraries
import pandas as pd
import numpy as np

# evaluation
---------------------
At this stage in the project you have built a model (or models) that appears to have high quality from a data analysis perspective. Before proceeding to final deployment of the model, it is important to more thoroughly evaluate the model and review the steps executed to construct the model to be certain it properly achieves the business objectives. A key objective is to determine if there is some important business issue that has not been sufficiently considered. At the end of this phase, a decision on the use of the data mining results should be reached.

## evaluate result
----------

### task

Previous evaluation steps dealt with factors such as the accuracy and generality of the model. This step assesses the degree to which the model meets the business objectives and seeks to determine if there is some business reason why this model is deficient. Another option of evaluation is to test the model(s) on test applications in the real application if time and budget constraints permit.

Moreover, evaluation also assesses other data mining results generated. Data mining results cover models which are necessarily related to the original business objectives and all other findings which are not necessarily related to the original business objectives but might also unveil additional challenges, information or hints for future directions.

### output

#### assessment of data mining results with respect to business success criteria 

Summarize assessment results in terms of business success criteria including a final statement whether the project already meets the initial business objectives.

#### approved models 

After model assessment with respect to business success criteria, the gen-erated models that meet the selected criteria become approved models.


## review process
------------------

### task

At this point the resultant model hopefully appears to be satisfactory and to satisfy business needs. It is now appropriate to do a more thorough review of the data mining engagement in order to determine if there is any important factor or task that has somehow been overlooked. This review also covers quality assurance issues, e.g., did we correctly build the model? Did we only use attributes that we are allowed to use and that are available for future analyses?

### output

Summarize the process review and highlight activities that have been missed and/or should be repeated. 


## determine next steps
----------

### task

According to the assessment results and the process review, the project decides how to proceed at this stage. The project needs to decide whether to finish this project and move on to deployment if appropriate or whether to initiate further iterations or set up new data mining projects. This task includes analyses of remaining resources and budget that influences the decisions.

### output

#### list of possible actions 

List the potential further actions along with the reasons for and against each option.

#### Decision 

Describe the decision as to how to proceed along with the rationale.


## note/questions
-------------

#### evaluate model
As we can see by coefficients, some countries has them as large as 20. It means customers from these countries are less likely to buy based on price and quantity(Brazil, Denmark, Greece, Japan, Lebanon, Lithuania), but more likely to buy just because they want to buy product. So we can up the price a little bit. \
These countries:
 - Austria
 - Bahrain
 - Belgium
 - Canada
 - Channel Islands
 - Cyprus
 - Czech Republic
 - European Community
 - France
 - Germany
 - Iceland
 - Israel
 - Italy
 - Malta
 - Netherlands
 - Norway
 - Poland
 - Portugal
 - RSA
 - Saudi Arabia
 - Spain
 - Sweden
 - USA
 - United Kingdom
 
have negative coefficient, it means they will buy looking at price and quantity of a product, and we need to do discounts, so they will buy more.\
Other countries have reasonable deviations and market can be played with using discounts.

Also best time for selling is 12:00. Taking it into consideration we can add ads with food(12:00 is lunch time), so that people will buy it with products. \
Average selling quantity is 4, average selling price is 1.6, we need to add more products with that price and around that quantity. \
If country coefficient is large, it means product is more likely to be bought in that country without looking at another characteristics.
#### review process
Best characteristics were found. 
Other algorithms, beside regression can be used to predict 'sum' column.
#### determine next step
Project needs to initiate further iterations, based on another data.
