==> s1/6_deployment.ipynb <==
# import libraries
import pandas as pd
import numpy as np

# deployment
---------------------

Creation of the model is generally not the end of the project. Even if the purpose of the model is to increase knowledge of the data, the knowledge gained will need to be organized and presented in a way that the customer can use it. It often involves applying "live" models within an organization's decision making processes, for example in real-time personalization of Web pages or repeated scoring of marketing databases. However, depending on the requirements, the deployment phase can be as simple as generating a report or as complex as implementing a repeatable data mining process across the enterprise. In many cases it is the customer, not the data analyst, who carries out the deployment steps. However, even if the analyst will not carry out the deployment effort it is important for the customer to understand up front what actions need to be carried out in order to actually make use of the created models.


## plan deployment
----------

### task

In order to deploy the data mining result(s) into the business, this task takes the evaluation results and concludes a strategy for deployment. If a general procedure has been identified to create the relevant model(s), this procedure is documented here for later deployment.

### output

Summarize deployment strategy including necessary steps and how to perform them.

## plan monitoring & maintenance
------------------

### task

Monitoring and maintenance are important issues if the data mining result becomes part of the day-to-day business and its environment. A careful preparation of a maintenance strategy helps to avoid unnecessarily long periods of incorrect usage of data mining results. In order to monitor the deployment of the data mining result(s), the project needs a detailed plan on the monitoring process. This plan takes into account the specific type of deployment.

### output

Summarize monitoring and maintenance strategy including necessary steps and how to perform them.


## produce final report
----------

### task

At the end of the project, the project leader and his team write up a final report. Depending on the deployment plan, this report may be only a summary of the project and its experiences (if they have not already been documented as an ongoing activity) or it may be a final and comprehensive presentation of the data mining result(s)

### output

#### final report 

This is the final written report of the data mining engagement. It includes all of the previous deliverables and summarizes and organizes the results.

#### final presentation 

There will also often be a meeting at the conclusion of the project where the results are verbally presented to the customer.




# review project
-------------

## task

Assess what went right and what went wrong, what was done well and what needs to be improved.

## output

Summarize important experiences made during the project. For example, pitfalls, misleading approaches or hints for selecting the best suited data mining techniques in similar situations could be part of this documentation. In ideal projects, experience documentation covers also any reports that have been written by individual project members during the project phases and their tasks.



## note/questions
-------------

#### plan deployment

#### plan monitoring & maintenance

#### produce final report

#### review project
